requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.25
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
email-validator>=2.0.0
python-multipart>=0.0.9
openai>=1.44.0
